agent_service/requirements.txt
google-adk>=0.1.0
google-genai>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
tensorflow>=2.15.0
joblib>=1.3.0
yfinance>=0.2.36
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
